Updates to parallel patterns.
